removed product quantity and implemented availability

+++ app/views/subviews/products.php
@@ -14,15 +14,24 @@
                     <a class=\"btn btn-primary\" href='/Product/delete/$product->product_id' onclick='return confirm(\"Product successfully deleted\")' class='m-2'>Delete</a></div>
                     </div>";
 
-            } else if (isset($_SESSION['order->store_id']) && $_SESSION['order->store_id'] != $product->store_id) {
+            } else if ($product->product_availability == 1 && isset($_SESSION['order->store_id']) && $_SESSION['order->store_id'] != $product->store_id) {
+ 
                     echo "<a class=\"btn btn-primary\" href='/Cart/addToCart/$product->product_id/$product->store_id' onclick='return confirm(\"Your current cart will be cleared if you proceed. Are you sure?\")' class='m-2'> Add to Cart </a> 
                         </div>
                         </div>";
             } else {
-                echo "<a class=\"btn btn-primary\" href='/Cart/addToCart/$product->product_id/$product->store_id' onclick='return confirm(\"Product successfully added to cart!\")' class='m-2'> Add to Cart </a> 
+                if ($product->product_availability == 0) {
+                    echo "<a class=\"btn btn-primary\" onclick='return confirm(\"Product is not available at the moment.\")' class='m-2'> Add to Cart </a> 
 
                     </div>
                     </div>";
+                }
+                else {
+                    echo "<a class=\"btn btn-primary\" href='/Cart/addToCart/$product->product_id/$product->store_id' onclick='return confirm(\"Product successfully added to cart!\")' class='m-2'> Add to Cart </a> 
+
+                        </div>
+                        </div>";
+                }
             }
         }
     ?>
